Sorting visualization complete
