+++ Visualizations/main.py
@@ -1,0 +1,11 @@
+from visualizations import animate_sort
+
+
+def main():
+    # Available sorting algorithms:
+    # 1. Bubble Sort
+    animate_sort("Bubble Sort")
+
+
+if __name__ == '__main__':
+    main()
